--HG--
extra : convert_revision : 56c2205cdbb9af64c30b381a80b4d14c97841da7

+++ src/arch/sparc/types.hh
@@ -59,7 +59,7 @@ namespace SparcISA
 
     typedef int RegContextVal;
 
-    typedef uint8_t RegIndex;
+    typedef uint16_t RegIndex;
 }
 
 #endif
